Added new landing image and renamed landing files
