+++ front-end/src/components/landing/landing-image.js
@@ -4,25 +4,12 @@ import SplitText from 'react-pose-text';
 import { Link } from 'react-router-dom';
 
 // styles
-import './landingImage.scss';
+import './landing-image.scss';
 
 // images
-import barChartBW from '../../../assets/bar-chart-bw.svg';
-import barChartOrig from '../../../assets/bar-chart-orig.svg'
-// import keyboard from  '../../../assets/keyboard.svg';
-// import keyboardHighlighted from '../../../assets/keyboard-highlighted.svg';
-import keyboard from  '../../../assets/hand-orig.svg';
-import keyboardHighlighted from '../../../assets/hand-highlighted.svg';
-import puzzleSolo from '../../../assets/puzzle-solo.png';
-
 
 // routes
 import * as routes from '../../routes';
-
-const ArrowDown = () => {
-  return <svg id='i-chevron-bottom' viewBox='0 0 32 32' width='32' height='32' fill='none' stroke='currentcolor' strokeLinecap='round' strokeLinejoin='round' strokeWidth='2'>
-    <path d='M30 12 L16 24 2 12' /></svg>;
-};
 
 // Animation properties for React Pose Text to add animation for each letter
 const charPoses = {
@@ -41,8 +28,8 @@ class LandingImage extends React.Component {
       onHover: {
         barChart: false,
         keyboard: false
-      }
-    }
+      },
+    };
   }
 
   onImageHover(event) {
@@ -75,18 +62,8 @@ class LandingImage extends React.Component {
               </div>
             </section>
             <div className='intro-headline'>
-              { /* Link to={routes.DASHBOARD_FRONTEND}>
-                {<img src={onHover.keyboard ? keyboardHighlighted : keyboard }
-                  onMouseOver={this.onImageHover.bind(this)}
-                   onMouseOut={this.onImageExit.bind(this)} alt='bar-chart' className='keyboard-image'/>
-               </Link> */ }
               { onHover.keyboard ? <h1 className='infoTextVolunteer'>I want to volunteer.</h1> : null }
               { onHover.barChart ? <h1 className='infoTextBusiness'>I am a business.</h1> : null }
-              { /* <Link to={routes.DASHBOARD_FRONTEND}>
-                <img src={onHover.barChart ? barChartOrig : barChartBW}
-                  onMouseOver={this.onImageHover.bind(this)}
-                     onMouseOut={this.onImageExit.bind(this)} alt='bar-chart' className='barChart-image'/>
-                 </Link> */ }
             </div>
           </div>
         </div>
