+++ front-end/src/components/landing/landing-image.scss
@@ -27,7 +27,7 @@
 }
 
 .landing-image {
-  background: url('../../../assets/landingimage.svg') no-repeat content-box;
+  background: url('../../../assets/landing-image.jpg') no-repeat content-box;
   background-size: cover;
   width: 100%;
   max-height: 28vw;
