+++ front-end/src/components/project-details/project-info.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Link from 'react-router';
+import { Link } from 'react-router-dom';
 import './project-info.scss';
 
 const ProjectInfo = (props) => {
